Start moving tests into a pair of central modules
Affects https://github.com/io7m/android/issues/1
Affects https://github.com/AULFA/android/issues/1
