--- simplified-rfc3339-tests-android/build.gradle
@@ -1,24 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-  compileSdkVersion 26
-  buildToolsVersion "26.0.2"
-
-  defaultConfig {
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-  }
-
-  testOptions {
-    execution 'ANDROID_TEST_ORCHESTRATOR'
-  }
-}
-
-description = 'simplified-rfc3339-tests-android'
-
-dependencies {
-  compile project(':simplified-rfc3339-tests')
-
-  compile 'com.android.support.test:runner:1.0.1'
-  androidTestImplementation 'com.android.support.test:runner:1.0.1'
-  androidTestUtil 'com.android.support.test:orchestrator:1.0.1'
-}
