--- simplified-rfc3339-tests-android/src/main/java/org/nypl/simplified/rfc3339/tests/android/RFC3339FormatterTest.java
@@ -1,18 +1,0 @@
-package org.nypl.simplified.rfc3339.tests.android;
-
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.runner.RunWith;
-import org.nypl.simplified.rfc3339.tests.RFC3339FormatterContract;
-
-/**
- * Android unit test frontend.
- */
-
-@RunWith(AndroidJUnit4.class)
-@SmallTest
-public final class RFC3339FormatterTest extends RFC3339FormatterContract
-{
-
-}
