--- simplified-rfc3339-tests-android/src/main/java/org/nypl/simplified/rfc3339/tests/android/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Android frontend for test suite.
- */
-
-@com.io7m.jnull.NonNullByDefault package org.nypl.simplified.rfc3339.tests.android;
-
