--- simplified-rfc3339-tests/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-  compileSdkVersion 26
-  buildToolsVersion "26.0.2"
-}
-
-description = 'simplified-rfc3339-tests'
-
-dependencies {
-  compile project(':simplified-rfc3339-core')
-  compile 'junit:junit:4.12'
-  testCompile 'junit:junit:4.12'
-}
