--- simplified-rfc3339-tests/src/main/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  package="org.nypl.simplified.rfc3339.tests">
-    <uses-sdk
-      android:minSdkVersion="19"
-      android:targetSdkVersion="25"/>
-
-</manifest>
