--- simplified-rfc3339-tests/src/main/java/org/nypl/simplified/rfc3339/tests/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Unit tests specified in a test-suite independent manner.
- */
-
-@com.io7m.jnull.NonNullByDefault package org.nypl.simplified.rfc3339.tests;
-
