--- simplified-rfc3339-tests/src/test/java/org/nypl/simplified/rfc3339/tests/local/RFC3339FormatterTest.java
@@ -1,11 +1,0 @@
-package org.nypl.simplified.rfc3339.tests.local;
-
-import org.nypl.simplified.rfc3339.tests.RFC3339FormatterContract;
-
-/**
- * JUnit4 test case frontend.
- */
-
-public final class RFC3339FormatterTest extends RFC3339FormatterContract {
-
-}
