--- simplified-rfc3339-tests/src/test/java/org/nypl/simplified/rfc3339/tests/local/package-info.java
@@ -1,6 +1,0 @@
-/**
- * JUnit 4 frontend for test suite.
- */
-
-@com.io7m.jnull.NonNullByDefault package org.nypl.simplified.rfc3339.tests.local;
-
