+++ simplified-tests-android/build.gradle
@@ -1,0 +1,24 @@
+apply plugin: 'com.android.library'
+
+android {
+  compileSdkVersion 26
+  buildToolsVersion "26.0.2"
+
+  defaultConfig {
+    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+  }
+
+  testOptions {
+    execution 'ANDROID_TEST_ORCHESTRATOR'
+  }
+}
+
+description = 'simplified-tests-android'
+
+dependencies {
+  compile project(':simplified-tests')
+
+  compile 'com.android.support.test:runner:1.0.1'
+  androidTestImplementation 'com.android.support.test:runner:1.0.1'
+  androidTestUtil 'com.android.support.test:orchestrator:1.0.1'
+}
