+++ simplified-tests-android/src/main/java/org/nypl/simplified/tests/android/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Android unit test frontend.
+ */
+
+package org.nypl.simplified.tests.android;
