+++ simplified-tests-android/src/main/java/org/nypl/simplified/tests/android/rfc3339/RFC3339FormatterTest.java
@@ -1,0 +1,18 @@
+package org.nypl.simplified.tests.android.rfc3339;
+
+import android.support.test.filters.SmallTest;
+import android.support.test.runner.AndroidJUnit4;
+
+import org.junit.runner.RunWith;
+import org.nypl.simplified.tests.rfc3339.RFC3339FormatterContract;
+
+/**
+ * Android unit test frontend.
+ */
+
+@RunWith(AndroidJUnit4.class)
+@SmallTest
+public final class RFC3339FormatterTest extends RFC3339FormatterContract
+{
+
+}
