+++ simplified-tests-android/src/main/java/org/nypl/simplified/tests/android/rfc3339/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Android frontend for test suite.
+ */
+
+@com.io7m.jnull.NonNullByDefault package org.nypl.simplified.tests.android.rfc3339;
+
