+++ simplified-tests/build.gradle
@@ -1,0 +1,15 @@
+apply plugin: 'com.android.library'
+
+android {
+  compileSdkVersion 26
+  buildToolsVersion "26.0.2"
+}
+
+description = 'simplified-tests'
+
+dependencies {
+  compile project(':simplified-rfc3339-core')
+
+  compile 'junit:junit:4.12'
+  testCompile 'junit:junit:4.12'
+}
