+++ simplified-tests/src/main/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  package="org.nypl.simplified.tests">
+    <uses-sdk
+      android:minSdkVersion="19"
+      android:targetSdkVersion="25"/>
+
+</manifest>
