+++ simplified-tests/src/main/java/org/nypl/simplified/tests/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Unit tests
+ */
+
+package org.nypl.simplified.tests;
+
