+++ simplified-tests/src/main/java/org/nypl/simplified/tests/rfc3339/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * RFC3339 unit tests.
+ */
+
+package org.nypl.simplified.tests.rfc3339;
+
