+++ simplified-tests/src/test/java/org/nypl/simplified/tests/local/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Local unit test frontend.
+ */
+
+package org.nypl.simplified.tests.local;
+
