+++ simplified-tests/src/test/java/org/nypl/simplified/tests/local/rfc3339/RFC3339FormatterTest.java
@@ -1,0 +1,11 @@
+package org.nypl.simplified.tests.local.rfc3339;
+
+import org.nypl.simplified.tests.rfc3339.RFC3339FormatterContract;
+
+/**
+ * JUnit4 test case frontend.
+ */
+
+public final class RFC3339FormatterTest extends RFC3339FormatterContract {
+
+}
