+++ simplified-tests/src/test/java/org/nypl/simplified/tests/local/rfc3339/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * RFC3339 local test frontend.
+ */
+
+package org.nypl.simplified.tests.local.rfc3339;
+
