+++ simplified-tests/src/test/resources/logback.xml
